requirements.txt
Pillow >= 10.3
click
cogapp
